+++ TunnelKitTests/DataPathPerformanceTests.swift
@@ -21,7 +21,7 @@ class DataPathPerformanceTests: XCTestCase {
         let ck = try! SecureRandom.safeData(length: 32)
         let hk = try! SecureRandom.safeData(length: 32)
         
-        let crypto = try! EncryptionProxy(.aes128cbc, .sha1, ck, ck, hk, hk)
+        let crypto = try! SessionProxy.EncryptionBridge(.aes128cbc, .sha1, ck, ck, hk, hk)
         encrypter = crypto.encrypter()
         decrypter = crypto.decrypter()
